Perf(utils): Move configured vocabulary functions to own modules for tree shaking
