+++ CHANGELOG.md
@@ -82,6 +82,7 @@ xx.xx.xxxx
 * **Enhancement** - `humanize` rescues common lowercase acronyms out of the box (`api_url` → `API URL`) through a `humanize.config` vocabulary seeded with `id`/`url`/`api`. Extend it once at app boot and every call inherits it, or pass a per-call `terms` map that layers over the global. `humanize.config` also carries the default `titleCase`/`dropId`/`constantCase`
 * **Feature** - Added a coercion suite — [`toBoolean`](https://next.semantic-ui.com/docs/api/utils/coercion#toboolean), [`toNumber`](https://next.semantic-ui.com/docs/api/utils/coercion#tonumber), [`toInteger`](https://next.semantic-ui.com/docs/api/utils/coercion#tointeger), [`toDate`](https://next.semantic-ui.com/docs/api/utils/coercion#todate), [`toString`](https://next.semantic-ui.com/docs/api/utils/coercion#tostring), with `coerceX` aliases — best-effort conversion of loose input to a target type. Each returns the type or `null` when there is no clean reading, never a poison value, so results compose with `??`. `toBoolean` reads a generous set of true/false spellings, `toNumber` never yields `NaN` or `Infinity`, `toDate` takes ISO strings and epoch-ms numbers only (ambiguous formats like `01/15/2024` return null rather than a guessed Date), and `toString` returns null for objects instead of `"[object Object]"`. `toBoolean` and `toString` take `{ loose: true }` to fall back to native truthiness or JSON rendering. Every helper takes `{ onInvalid: 'passthrough' }` to return the original value on a failed coercion instead of `null`, so a schema or validator sees the bad input rather than an erasing `null` (the passthrough return widens the TypeScript type to include the input). `toBoolean` reads its vocabulary and defaults from an editable `toBoolean.config` (mirroring `humanize.config`), so a locale or domain teaches it new true/false spellings once at boot instead of on every call
 * **Feature** - Added a bytes module — [`toBase64`](https://next.semantic-ui.com/docs/api/utils/bytes#tobase64) and [`fromBase64`](https://next.semantic-ui.com/docs/api/utils/bytes#frombase64) — unicode-safe base64 encode/decode, the pair `btoa`/`atob` never were. A string round-trips through its UTF-8 bytes (emoji and accents survive), `toBase64` also takes an `ArrayBuffer` or typed array and a `{ urlSafe }` option, and `fromBase64` decodes both alphabets, strips MIME/PEM line wrapping, and returns a string or `Uint8Array` — or `null` on malformed input, never a throw
+* **Enhancement** - `humanize`, `toTitleCase`, and `getArticle` live in their own modules so a bundle only pays for the vocabularies it imports — the attribute-codec path that pulls in string casing no longer carries their configs
 * **Feature** - Arbitrary vocabularies now live in editable configs, set once at app boot: [`toTitleCase.config.stopWords`](https://next.semantic-ui.com/docs/api/utils/strings#totitlecase) (style guides disagree on the list, `humanize`'s `titleCase` reads the same one), [`getArticle.config.exceptions`](https://next.semantic-ui.com/docs/api/utils/strings#getarticle) for sound-contradicts-spelling words (`an hour`, `a university` now correct out of the box), and [`formatDate.config.timezones`](https://next.semantic-ui.com/docs/api/utils/dates#formatdate) for remapping an ambiguous shorthand like `IST`. `toDate` reads unix-second timestamps (a JWT `exp`) via `{ epoch: 'seconds' }`
 * **Feature** - Added `unescapeHTML()` for converting HTML entities back to characters — the inverse of `escapeHTML`
 * **Bug** - Fixed `escapeHTML()` producing entities without semicolons (e.g. `&amp` instead of `&amp;`)
