+++ packages/utils/src/index.js
@@ -12,10 +12,13 @@ export * from './debug.js';
 export * from './environment.js';
 export * from './equality.js';
 export * from './functions.js';
+export * from './get-article.js';
 export * from './html.js';
+export * from './humanize.js';
 export * from './loops.js';
 export * from './numbers.js';
 export * from './objects.js';
 export * from './regexp.js';
 export * from './strings.js';
+export * from './to-title-case.js';
 export * from './types.js';
